snapshots: LVM related code to new file
Updates: #145
